refactor: drop Button — Link block, add Button — Icon block
Remove ButtonLinkButton (contradicts best practice about buttons vs
links). Add ButtonWithIcon showing leading icons across all 4 variants
with heroicons.

Made-with: Cursor
