--- packages/cli/templates/blocks/components/Button/ButtonLinkButton.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'Button — Link',
-  description: 'Buttons that navigate by rendering as anchor elements when href is provided. Use when you need button styling but the action is navigation.',
-  isReady: true,
-  aspectRatio: 16 / 9,
-  componentsUsed: ['Button', 'Layout', 'Text'],
-};
